% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malirrm-package.R
\docType{package}
\name{malirrm-package}
\alias{malirrm-package}
\alias{testduplicate-package}
\title{malirrm: Mali RRM}
\description{
RRM evaluation in Mali.
}
\author{
\strong{Maintainer}: Colin Walder \email{cwalder@ethz.ch} (\href{https://orcid.org/0009-0006-0969-1954}{ORCID})

}
\keyword{internal}
