% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malirrm.R
\docType{data}
\name{malirrm}
\alias{malirrm}
\title{Mali RRM

Mali RRM.}
\format{
A tibble with 74 rows and 10 variables.
\describe{
  \item{start}{ Beginning of survey.}
  \item{end}{ End of survey.}
  \item{date}{ Date.}
  \item{region}{ Region.}
  \item{cercle}{ Cercle.}
  \item{quartier}{ Quartier.}
  \item{marche}{ Market.}
  \item{consent}{ Consent.}
  \item{marches_produits}{ Market product.}
  \item{marches_produits_alimentaires}{ Market product alimentaires.}
}
}
\usage{
malirrm
}
\description{
Mali RRM

Mali RRM.
}
\keyword{datasets}
